use floats instead of doubles

+++ include/Boundary.h
@@ -12,7 +12,7 @@ public:
 
 	virtual void draw2D() = 0;
 	//virtual void draw3D() = 0;
-	virtual double distance(const Particle &particle) const = 0;
+	virtual float distance(const Particle &particle) const = 0;
 	virtual Vec3d getNormal(const Particle &particle) const = 0;
 
 	bool isTemporary() const {
