+++ include/Vec3d.h
@@ -7,11 +7,11 @@
 
 struct Vec3d {
 
-	double x;
-	double y;
-	double z;
+	float x;
+	float y;
+	float z;
 
-	Vec3d(double x, double y, double z) :
+	Vec3d(float x, float y, float z) :
 			x(x), y(y), z(z) {
 	}
 
@@ -32,11 +32,11 @@ inline Vec3d Substract(const Vec3d &a, const Vec3d &b) {
 	return Vec3d(a.x - b.x, a.y - b.y, a.z - b.z);
 }
 
-inline Vec3d Multiply(const Vec3d &a, const double m) {
+inline Vec3d Multiply(const Vec3d &a, const float m) {
 	return Vec3d(a.x * m, a.y * m, a.z * m);
 }
 
-inline Vec3d Divide(const Vec3d &a, const double m) {
+inline Vec3d Divide(const Vec3d &a, const float m) {
 	// TODO: handle division by zero
 	return Vec3d(a.x / m, a.y / m, a.z / m);
 }
@@ -49,15 +49,15 @@ inline Vec3d operator-(const Vec3d &a, const Vec3d &b) {
 	return Substract(a, b);
 }
 
-inline Vec3d operator*(const double m, const Vec3d &a) {
+inline Vec3d operator*(const float m, const Vec3d &a) {
 	return Multiply(a, m);
 }
 
-inline Vec3d operator*(const Vec3d &a, const double m) {
+inline Vec3d operator*(const Vec3d &a, const float m) {
 	return Multiply(a, m);
 }
 
-inline double operator*(const Vec3d &a, const Vec3d &b) {
+inline float operator*(const Vec3d &a, const Vec3d &b) {
 	return a.x * b.x + a.y * b.y + a.z * b.z;
 }
 
@@ -66,16 +66,16 @@ inline Vec3d crossProduct(const Vec3d &a, const Vec3d &b) {
 			a.x * b.y - a.y * b.x);
 }
 
-inline Vec3d operator/(const Vec3d &a, const double m) {
+inline Vec3d operator/(const Vec3d &a, const float m) {
 	return Divide(a, m);
 }
 
-inline double abs(const Vec3d &a) {
+inline float abs(const Vec3d &a) {
 	return std::sqrt(a.x * a.x + a.y * a.y + a.z * a.z);
 }
 
 inline Vec3d norm(const Vec3d &a) {
-	double length = abs(a);
+	float length = abs(a);
 	if (length > SMALL) {
 		return a / length;
 	} else {
